made minor edits to variable names and formatting/spacing

+++ client/src/components/Signup.js
@@ -3,25 +3,24 @@ import { VStack } from '@chakra-ui/layout';
 import { FormControl, FormLabel } from '@chakra-ui/form-control';
 import { Input, InputGroup, InputRightElement } from '@chakra-ui/input';
 import { Button } from '@chakra-ui/button';
-import { useToast } from '@chakra-ui/react';
+import { useToast } from '@chakra-ui/toast';
 import axios from 'axios';
 import { useHistory } from 'react-router-dom';
 
 const Signup = () => {
     const [show, setShow] = useState(false);
+    const handleClick = () => setShow(!show);
+    const toast = useToast();
+    const history = useHistory();
     const [name, setName] = useState();
     const [email, setEmail] = useState();
     const [password, setPassword] = useState();
     const [confirmpassword, setConfirmpassword] = useState();
     const [pic, setPic] = useState();
-    const [loading, setLoading] = useState(false); 
-    const toast = useToast();
-    const history = useHistory();
-
-    const handleClick = () => setShow(!show);
-
+    const [picLoading, setPicLoading] = useState(false); 
+    
     const postDetails = (pics) => { 
-        setLoading(true);
+        setPicLoading(true);
         if (pics === undefined) {
             toast({
                 title: 'Please select an image',
@@ -32,7 +31,7 @@ const Signup = () => {
             });
             return;
         }
-
+        console.log(pics);
         if (pics.type === 'image/jpeg' || pics.type === 'image/png') {
             const data = new FormData();
             data.append('file', pics);
@@ -45,11 +44,11 @@ const Signup = () => {
                 .then((res) => res.json())
                 .then((data) => {
                     setPic(data.url.toString());
-                    setLoading(false);
+                    setPicLoading(false);
                 })
                 .catch((err) => {
                     console.log(err);
-                    setLoading(false);
+                    setPicLoading(false);
                 });
         } else {
             toast({
@@ -59,13 +58,13 @@ const Signup = () => {
                 isClosable: true,
                 position: 'bottom'
             });
-            setLoading(false);
+            setPicLoading(false);
             return;
         }
     }; 
 
     const submitHandler = async () => {
-        setLoading(true);
+        setPicLoading(true);
         if (!name || !email || !password || !confirmpassword) {
             toast({
                 title: 'Please fill out all fields',
@@ -74,33 +73,46 @@ const Signup = () => {
                 isClosable: true,
                 position: 'bottom'
             });
-            setLoading(false);
+            setPicLoading(false);
             return;
         }
-        
+        if (password !== confirmpassword) {
+            toast({
+                title: 'Passwords do not match',
+                status: 'warning',
+                duration: 5000,
+                isClosable: true,
+                position: 'bottom'
+            });
+            return;
+        }
+        console.log(name, email, password, pic);
         try {
             const config = {
                 headers: {
-                    'Content-type': 'application/json',
-                },
+                    'Content-type': 'application/json'
+                }
             };
-
             const { data } = await axios.post(
                 '/api/user',
-                { username, email, password, pic },
+                {
+                    name,
+                    email,
+                    password,
+                    pic
+                },
                 config
             );
+            console.log(data);
             toast({
-                title: 'Registration Successful',
+                title: 'Registration successful!',
                 status: 'success',
                 duration: 5000,
                 isClosable: true,
                 position: 'bottom'
             });
-
             localStorage.setItem('userInfo', JSON.stringify(data));
-
-            setLoading(false);
+            setPicLoading(false);
             history.push('/chats')
         } catch (error) {
             toast({
@@ -111,6 +123,7 @@ const Signup = () => {
                 isClosable: true,
                 position: 'bottom'
             });
+            setPicLoading(false);
         }
     };
 
@@ -174,7 +187,7 @@ const Signup = () => {
                 width='100%'
                 style={{ marginTop: 15 }}
                 onClick={submitHandler}
-                isLoading={loading}
+                isLoading={picLoading}
             >
                 Sign Up
             </Button>
